Move testing hook
