+++ src/client/components/MyArea/MyArea.hooks.test.ts
@@ -3,7 +3,7 @@ import { describe, expect, it, test, vi } from 'vitest';
 
 import { useMapLocations, useSetMapCenterAtLocation } from './MyArea.hooks';
 import { appStateAtom } from '../../hooks/useAppState';
-import { renderRecoilHook } from '../../utils/renderRecoilHook';
+import { renderRecoilHook } from '../../../testing/renderRecoilHook';
 
 const mapInstanceMock = {
   setView: vi.fn(),
